# lib/apiservers/portlayer/swagger.json
{
	"swagger": "2.0",
	"basePath": "/",
	"info": {
		"description": "Port Layer API",
		"title": "Port Layer API",
		"version": "v0.0.1"
	},
	"produces": [
		"application/json"
	],
	"consumes": [
		"application/json"
	],
	"schemes": [
		"http"
	],
	"paths": {
...
		"/archive": {
			"get": {
				"description": "Exports a tar archive from the target device.",
				"summary": "Export tar archive from supplied target",
				"tags": [
					"storage"
				],
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/x-gzip"
				],
				"operationId": "ExportArchive",
				"parameters": [
					{
						"name":"objectID",
						"in":"query",
						"required": true,
						"type" : "string",
						"x-nullable": false
					},
					{
						"name":"targetPath",
						"in":"query",
						"required": true,
						"type": "string",
						"x-nullable": false
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"format":"binary"
						}
					},
					"404" : {
						"description": "Supplied target not found",
						"schema": {
							"$ref": "#/definitions/Error"
						}
					},
					"409": {
						"description": "Device or resource has an unresolvable conflict",
						"schema": {
							"$ref": "#/definitions/Error"
						}
					},
					"500": {
						"description": "failed to export tar archive from target"
					},
					"default": {
						"description": "error",
						"schema": {
							"$ref": "#/definitions/Error"
						}
					}
				}
			},
			"post": {
				"description": "Imports a tar stream to a device at the target path",
				"summary": "Import tar stream to supplied target",
				"tags": [
					"storage"
				],
				"consumes": [
					"application/x-tar"
				],
				"operationId": "ImportArchive",
				"parameters": [
					{
						"name":"objectID",
						"in":"query",
						"required": true,
						"type" : "string",
						"x-nullable": false
					},
					{
						"name":"destinationPath",
						"in":"query",
						"required": true,
						"type": "string",
						"x-nullable": false
					},
					{
						"name": "Archive",
						"required": true,
						"in": "body",
						"schema": {
							"type":"string",
							"format":"binary"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"404" : {
						"description": "Supplied target not found",
						"schema": {
							"$ref": "#/definitions/Error"
						}
					},
					"409": {
						"description": "Device or resource has an unresolvable conflict",
						"schema": {
							"$ref": "#/definitions/Error"
						}
					},
					"500": {
						"description": "failed to export tar archive from target"
					},
					"default": {
						"description": "error",
						"schema": {
							"$ref": "#/definitions/Error"
						}
					}
				}
			},
			"head": {
				"description": "Fetches filesystem stats on the device id at the specified path",
				"operationId": "StatPath",
				"tags": [
					"storage"
				],
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"name":"deviceID",
						"in":"query",
						"required": true,
						"type" : "string",
						"x-nullable": false
					},
					{
						"name":"targetPath",
						"in":"query",
						"required": true,
						"type": "string",
						"x-nullable": false
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"headers": {
							"name": {
								"type": "string",
								"x-nullable": false
							},
							"mode": {
								"type": "integer",
								"format": "uint32"
							},
							"size": {
								"type": "integer",
								"format": "int64"
							},
							"linkTarget": {
								"type": "string",
								"x-nullable": false
							}
						}
					},
					"404" : {
						"description": "supplied target not found"
					},
					"500": {
						"description": "failed to fetch stat path from target"
					},
					"default": {
						"description": "error"
					}
				}
			}
		},
		"/storage": {
			"post": {
				"description": "Creates a location to store images",
				"summary": "creates an image store",
				"tags": [
					"storage"
				],
				"operationId": "CreateImageStore",
				"parameters": [
					{
						"name": "body",
						"in": "body",
						"schema": {
							"$ref": "#/definitions/ImageStore"
						}
					}
				],
				"responses": {
					"201": {
						"description": "Created",
						"schema": {
							"$ref": "#/definitions/StoreUrl"
						}
					},
					"409": {
						"description": "An image store with that name already exists.",
						"schema": {
							"$ref": "#/definitions/Error"
						}
					},
					"default": {
						"description": "error",
						"schema": {
							"$ref": "#/definitions/Error"
						}
					}
				}
			}
		},
		"/storage/{store_name}": {
			"get": {
				"description": "Retrieves a list of images given a list of image IDs, or all images in the image store if no param is passed.",
				"summary": "Retrieve a list of images in an image store",
				"tags": [
					"storage"
				],
				"operationId": "ListImages",
				"parameters": [
					{
						"name": "store_name",
						"type": "string",
						"in": "path",
						"required": true
					},
					{
						"name": "ids",
						"type": "array",
						"in": "query",
						"items": {
							"type": "string",
							"collectionFormat": "csv"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/Image"
							}
						}
					},
					"404": {
						"description": "Not found",
						"schema": {
							"$ref": "#/definitions/Error"
						}
					},
					"default": {
						"description": "error",
						"schema": {
							"$ref": "#/definitions/Error"
						}
					}
				}
			},
			"post": {
				"description": "Creates a new image layer in an image store",
				"summary": "Creates a new image layer",
				"tags": [
					"storage"
				],
				"operationId": "WriteImage",
				"consumes": [
					"application/octet-stream"
				],
				"parameters": [
					{
						"name": "image_file",
						"in": "body",
						"schema": {
							"type": "string",
							"format": "binary"
						}
					},
					{
						"name": "store_name",
						"type": "string",
						"in": "path",
						"required": true
					},
					{
						"name": "image_id",
						"type": "string",
						"in": "query",
						"required": true
					},
					{
						"name": "parent_id",
						"type": "string",
						"in": "query",
						"required": true
					},
					{
						"name": "sum",
						"type": "string",
						"in": "query",
						"required": true
					},
					{
						"name": "metadatakey",
						"type": "string",
						"x-nullable": false,
						"in": "query"
					},
					{
						"name": "metadataval",
						"type": "string",
						"x-nullable": false,
						"in": "query"
					}
				],
				"responses": {
					"201": {
						"description": "Created",
						"schema": {
							"$ref": "#/definitions/Image"
						}
					},
					"default": {
						"description": "error",
						"schema": {
							"$ref": "#/definitions/Error"
						}
					}
				}
			}
		},
		"/storage/{store_name}/info/{id}": {
			"get": {
				"description": "Inspect an image by id in an image store",
				"summary": "Inspect an image",
				"tags": [
					"storage"
				],
				"operationId": "GetImage",
				"parameters": [
					{
						"name": "store_name",
						"type": "string",
						"in": "path",
						"required": true
					},
					{
						"name": "id",
						"type": "string",
						"in": "path",
						"required": true
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/Image"
						}
					},
					"404": {
						"description": "Not found",
						"schema": {
							"$ref": "#/definitions/Error"
						}
					},
					"default": {
						"description": "error",
						"schema": {
							"$ref": "#/definitions/Error"
						}
					}
				}
			},
			"delete": {
				"description": "Delete an image by id in an image store",
				"summary": "Delete an image",
				"tags": [
					"storage"
				],
				"operationId": "DeleteImage",
				"parameters": [
					{
						"name": "store_name",
						"type": "string",
						"in": "path",
						"required": true
					},
					{
						"name": "id",
						"type": "string",
						"in": "path",
						"required": true
					},
					{
						"name": "keepNodes",
						"type": "array",
						"required": true,
						"in": "query",
						"items": {
							"type": "string",
							"collectionFormat": "csv"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/Image"
							}
						}
					},
					"404": {
						"description": "Not found",
						"schema": {
							"$ref": "#/definitions/Error"
						}
					},
					"423": {
						"description": "In use",
						"schema": {
							"$ref": "#/definitions/Error"
						}
					},
					"default": {
						"description": "error",
						"schema": {
							"$ref": "#/definitions/Error"
						}
					}
				}
			}
		},
		"/storage/{store_name}/tar/{id}": {
			"get": {
				"description": "Get an image by id in an image store as a tar file",
				"summary": "Get an image as a tar file",
				"tags": [
					"storage"
				],
				"operationId": "GetImageTar",
				"parameters": [
					{
						"name": "store_name",
						"type": "string",
						"in": "path",
						"required": true
					},
					{
						"name": "id",
						"type": "string",
						"in": "path",
						"required": true
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string",
							"format": "binary"
						}
					},
					"404": {
						"description": "Not found"
					},
					"default": {
						"description": "error",
						"schema": {
							"$ref": "#/definitions/Error"
						}
					}
				}
			}
		},
		"/storage/volumestores": {
			"get": {
				"description": "Get a list of available volume store locations",
				"operationId": "VolumeStoresList",
				"tags": [
					"storage"
				],
				"consumes": [
					"application/json",
					"application/octet-stream"
				],
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/VolumeStoresListResponse"
						}
					},
					"404": {
						"description": "Not found",
						"schema": {
							"$ref": "#/definitions/Error"
						}
					},
					"500": {
						"description": "error",
						"schema": {
							"$ref": "#/definitions/Error"
						}
					}
				}
			}
		},
...
	},
	"definitions": {
		"Error": {
			"type": "object",
			"required": [
				"message"
			],
			"properties": {
				"code": {
					"type": "integer",
					"format": "int64"
				},
				"message": {
					"type": "string",
					"x-nullable": false
				}
			}
		},
...
		"StoreUrl": {
			"type": "object",
			"required": [
				"url"
			],
			"properties": {
				"code": {
					"type": "integer",
					"format": "int64"
				},
				"url": {
					"x-nullable": false,
					"type": "string"
				}
			}
		},
		"ImageStore": {
			"type": "object",
			"required": [
				"name"
			],
			"properties": {
				"name": {
					"type": "string",
					"x-nullable": false
				}
			}
		},
		"Image": {
			"type": "object",
			"required": [
				"ID",
				"Store"
			],
			"properties": {
				"ID": {
					"x-nullable": false,
					"type": "string"
				},
				"Store": {
					"x-nullable": false,
					"type": "string"
				},
				"SelfLink": {
					"type": "string"
				},
				"Parent": {
					"type": "string"
				},
				"Metadata": {
					"type": "object",
					"additionalProperties": {
						"type": "string"
					}
				}
			}
		},
...
		"VolumeStoresListResponse": {
			"type": "object",
			"required": [
				"Stores"
			],
			"properties": {
				"Stores": {
					"type": "array",
					"items": {
						"type": "string"
					}
				}
			}
		},
...
	}
}